requirements.txt
Django==1.8.4
django-appconf==1.0.1
django-braces==1.8.0
django-debug-toolbar==1.3.2
django-sql-explorer==0.7
django-widget-tweaks==1.4.1
flake8==2.5.4
gunicorn==19.3.0
hashids==1.1.0
mccabe==0.4.0
pep8==1.7.0
psycopg2==2.6.1
pyflakes==1.0.0
python-dateutil==2.4.2
six==1.9.0
sqlparse==0.1.15
wheel==0.24.0
